PersonFactory: combine two words to make test callsigns (#1359)
I'm pretty sure this fixes the flakiness we've been seeing
in TimesheetControllerTest and AccessDocumentControllerTest.
I suspect that we're creating Persons with duplicate callsigns,
violating the constraint on the Person table. This change makes the
tests totally reliable on my desktop (while I can easily
reproduce the flakiness without the change), and without
downsides, this seems worth merging.