Factored out cache operations into ClubhouseCache.
The laravel scheduler wants a persistent external cache like a database or Redis to prevent duplicate task executions. There appears not to be a way to configure the scheduler to use a different cache store other than the default. Because of this, the Clubhouse has use an explicitly defined store (Octane -- a super fast in-app ephemeral cache). The default cache cannot be set to Octane. Sigh.